Language score as int

+++ trigger/models/Language.py
@@ -3,4 +3,4 @@ from typing import NamedTuple
 
 class Language(NamedTuple):
     name: str
-    score: float
+    score: int
